+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### 12.0
  - Added 6Ghz partial support (some scans, some attacks, capable adapter detection, etc.)
 
+### 11.41
+ - Added a function to simplify hooking language strings from plugins
+ - Enterprise identities capture feature
+ - Fixed broken dependency menu validation
+
 ### 11.40
  - Code refactor of the heredocs used for the Evil Twin, WEP and WPS attacks
  - Standardize xterm parameters and adjust some colors to override the system's default settings